refactore: move to screen
